Through all videos

+++ src/features/comment/actions/delete-comment.ts
@@ -1,32 +1,35 @@
-'use server';
+"use server";
 
-import { revalidatePath } from 'next/cache';
+import { revalidatePath } from "next/cache";
 import {
-  fromErrorToFormState,
-  toFormState,
-} from '@/components/form/utils/to-form-state';
-import { getCurrentAuthOrRedirect } from '@/features/auth/queries/get-current-auth-or-redirect';
-import * as ticketService from '@/features/ticket/services';
-import { prisma } from '@/lib/prisma';
-import { ticketPath } from '@/paths';
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
+import { prisma } from "@/lib/prisma";
+import { ticketPath } from "@/paths";
 
 export const deleteComment = async (id: string) => {
-  const { user } = await getCurrentAuthOrRedirect();
+  const { user } = await getAuthOrRedirect();
 
-  try {
-    const comment = await prisma.comment.delete({
-      where: {
-        id,
-        userId: user.id,
-      },
-    });
+  const comment = await prisma.comment.findUnique({
+    where: { id },
+  });
 
-    await ticketService.disconnectReferencedTickets(comment);
-  } catch (error) {
-    return fromErrorToFormState(error);
+  if (!comment || !isOwner(user, comment)) {
+    return toActionState("ERROR", "Not authorized");
   }
 
-  revalidatePath(ticketPath(id));
+  try {
+    await prisma.comment.delete({
+      where: { id },
+    });
+  } catch (error) {
+    return fromErrorToActionState(error);
+  }
 
-  return toFormState('SUCCESS', 'Comment deleted');
+  revalidatePath(ticketPath(comment.ticketId));
+
+  return toActionState("SUCCESS", "Comment deleted");
 };
