+++ src/features/ticket/actions/delete-ticket.ts
@@ -1,28 +1,41 @@
-'use server';
+"use server";
 
-import { revalidatePath } from 'next/cache';
-import { redirect } from 'next/navigation';
-import { setCookieByKey } from '@/actions/cookies';
-import { fromErrorToFormState } from '@/components/form/utils/to-form-state';
-import { getCurrentAuthOrRedirect } from '@/features/auth/queries/get-current-auth-or-redirect';
-import { prisma } from '@/lib/prisma';
-import { ticketsPath } from '@/paths';
+import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
+import { setCookieByKey } from "@/actions/cookies";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
+import { prisma } from "@/lib/prisma";
+import { ticketsPath } from "@/paths";
 
 export const deleteTicket = async (id: string) => {
-  const { user } = await getCurrentAuthOrRedirect();
+  const { user } = await getAuthOrRedirect();
 
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not authorized");
+    }
+
     await prisma.ticket.delete({
       where: {
         id,
-        userId: user.id,
       },
     });
   } catch (error) {
-    return fromErrorToFormState(error);
+    return fromErrorToActionState(error);
   }
 
   revalidatePath(ticketsPath());
-  setCookieByKey('toast', 'Ticket deleted');
+  setCookieByKey("toast", "Ticket deleted");
   redirect(ticketsPath());
 };
